Reorganizing components folder
