+++ src/components/modals/NewFeaturesModal.vue
